Rename widgets/{multichoice->multi_choice}.ts
