+++ bokehjs/src/lib/models/widgets/index.ts
@@ -24,7 +24,7 @@ export {Markup}              from "./markup"
 export {MultiSelect}         from "./multiselect"
 export {Paragraph}           from "./paragraph"
 export {PasswordInput}       from "./password_input"
-export {MultiChoice}         from "./multichoice"
+export {MultiChoice}         from "./multi_choice"
 export {NumericInput}        from "./numeric_input"
 export {PreText}             from "./pretext"
 export {RadioButtonGroup}    from "./radio_button_group"
