+++ bokehjs/src/lib/models/widgets/multi_choice.ts
